+++ packages/tokens-studio-for-figma/src/plugin/sendDocumentChange.ts
@@ -1,7 +1,14 @@
 import { sendSelectionChange } from './sendSelectionChange';
+import { notifyInstancesCreated } from './notifiers';
 
 export async function sendDocumentChange(event: DocumentChangeEvent) {
   console.log('document change event', event);
+
+  // Check if event has documentChanges before processing
+  if (!event.documentChanges) {
+    return;
+  }
+
   if (event.documentChanges.length === 1 && event.documentChanges[0].type === 'PROPERTY_CHANGE' && event.documentChanges[0].id === '0:0') {
     return;
   }
@@ -10,4 +17,24 @@ export async function sendDocumentChange(event: DocumentChangeEvent) {
     return;
   }
   await sendSelectionChange();
+
+  // Look for CREATE events of instance nodes and notify UI
+  const instanceCreations = event.documentChanges.filter(
+    (change) => change.type === 'CREATE' && change.origin === 'LOCAL',
+  );
+
+  if (instanceCreations.length > 0) {
+    // Check if any of the created nodes are instance nodes
+    const createdInstanceNodes = instanceCreations
+      .map((change) => figma.getNodeById(change.id))
+      .filter((node) => node && node.type === 'INSTANCE');
+
+    if (createdInstanceNodes.length > 0) {
+      // Set selection to the new instance nodes for potential token application
+      figma.currentPage.selection = createdInstanceNodes as SceneNode[];
+
+      // Notify UI that instances were created - UI will decide if it should apply tokens
+      notifyInstancesCreated(createdInstanceNodes.length);
+    }
+  }
 }
